refactor lstm ctest with dropout pattern test enabled
